+++ src/app/connexion/confirmation-email.tsx
@@ -25,19 +25,23 @@ const ConfirmationEmail = () => {
 
   return (
     <SafeAreaView style={{ flex: 1 }}>
-      <ScrollView className="h-1 bg-[#0594ae]">
+      <ScrollView className="h-1 bg-custom-background">
         <TopBarBack title={t('confirmation_email')} />
         <VStack className="px-10 justify-center items-center">
           <HStack className="justify-center">
             <MaterialCommunityIcons
               name="email-check-outline"
               size={128}
-              color="white"
+              className="text-custom-bg-inverse"
             />
           </HStack>
-          <Text className="text-white">{t('confirmation_email_text_1')}</Text>
-          <Text className="text-white">{email}</Text>
-          <Text className="text-white">{t('confirmation_email_text_2')}</Text>
+          <Text className="text-typography-white">
+            {t('confirmation_email_text_1')}
+          </Text>
+          <Text className="text-typography-white">{email}</Text>
+          <Text className="text-typography-white">
+            {t('confirmation_email_text_2')}
+          </Text>
           <Button onPress={() => router.navigate('/')} className="mt-2">
             <ButtonText>{t('retour_accueil')}</ButtonText>
           </Button>