+++ src/app/connexion/index.tsx
@@ -70,7 +70,7 @@ const Authentification = () => {
 
   return (
     <SafeAreaView style={{ flex: 1 }}>
-      <ScrollView className="h-1 bg-[#0594ae]">
+      <ScrollView className="h-1 bg-custom-background">
         <TopBarBack title={t('authentification')} />
         <VStack className="flex-1 px-10 justify-between">
           <VStack className="mb-5">
@@ -79,14 +79,12 @@ const Authentification = () => {
               isRequired={true}
               className="mb-5"
             >
-              <FormControlLabel className="mb-1">
-                <FormControlLabelText className="text-white">
-                  {t('email')}
-                </FormControlLabelText>
+              <FormControlLabel className="mb-1 *:text-typography-white">
+                <FormControlLabelText>{t('email')}</FormControlLabelText>
               </FormControlLabel>
-              <Input>
+              <Input className="border-custom-bg-inverse">
                 <InputField
-                  className="text-white placeholder:text-white"
+                  className="text-typography-white placeholder:text-typography-white"
                   placeholder={t('email_adresse')}
                   keyboardType="email-address"
                   returnKeyType="next"
@@ -107,14 +105,12 @@ const Authentification = () => {
               isRequired={true}
               className="mb-5"
             >
-              <FormControlLabel className="mb-1">
-                <FormControlLabelText className="text-white">
-                  {t('mot_de_passe')}
-                </FormControlLabelText>
+              <FormControlLabel className="mb-1 *:text-typography-white">
+                <FormControlLabelText>{t('mot_de_passe')}</FormControlLabelText>
               </FormControlLabel>
-              <Input>
+              <Input className="border-custom-bg-inverse">
                 <InputField
-                  className="text-white placeholder:text-white"
+                  className="text-typography-white placeholder:text-typography-white"
                   placeholder={t('mot_de_passe')}
                   secureTextEntry={!showPassword}
                   autoCapitalize={'none'}
@@ -125,6 +121,7 @@ const Authentification = () => {
                   <FontAwesome5.Button
                     name={showPassword ? 'eye' : 'eye-slash'}
                     backgroundColor="#00000000"
+                    className="text-custom-bg-inverse"
                     iconStyle={{ marginRight: 0 }}
                     size={16}
                     onPress={() => setShowPassword(!showPassword)}
@@ -138,12 +135,12 @@ const Authentification = () => {
                 </FormControlErrorText>
               </FormControlError>
               <Button
-                className="text-white self-end"
+                className="text-typography-white self-end"
                 size="sm"
                 variant="link"
                 isDisabled={true}
               >
-                <ButtonText className="text-white">
+                <ButtonText className="text-typography-white">
                   {t('mot_de_passe_oublie')}
                 </ButtonText>
               </Button>
@@ -159,7 +156,7 @@ const Authentification = () => {
             </FormControl>
           </VStack>
           <VStack space="md">
-            <Text className="text-white self-center" size="lg">
+            <Text className="text-typography-white self-center" size="lg">
               {t('ou_connecter_avec')}
             </Text>
             <HStack className="flex" space="lg">
@@ -171,11 +168,11 @@ const Authentification = () => {
               >
                 <FontAwesome5
                   name="apple"
-                  color="white"
+                  className="text-custom-bg-inverse"
                   size={18}
                   style={{ marginRight: 5 }}
                 />
-                <ButtonText className="text-white">Apple</ButtonText>
+                <ButtonText className="text-typography-white">Apple</ButtonText>
               </Button>
               <Button
                 className="grow"
@@ -185,17 +182,18 @@ const Authentification = () => {
               >
                 <FontAwesome5
                   name="google"
-                  color="white"
                   size={14}
-                  className="mr-2"
+                  className="text-custom-bg-inverse mr-2"
                 />
-                <ButtonText className="text-white">Google</ButtonText>
+                <ButtonText className="text-typography-white">
+                  Google
+                </ButtonText>
               </Button>
             </HStack>
           </VStack>
           <Divider className="my-5" />
           <VStack space="md">
-            <Text className="text-white self-center" size="lg">
+            <Text className="text-typography-white self-center" size="lg">
               {t('pas_encore_compte')}
             </Text>
             <Button