+++ src/app/tournoi/(tabs)/matchs.tsx
@@ -7,11 +7,12 @@ import { useSelector } from 'react-redux';
 import ListeMatchs from '@components/matchs/liste-matchs';
 import { TypeTournoi } from '@/types/enums/typeTournoi';
 import { FontAwesome5 } from '@expo/vector-icons';
+import { StyledTopTabs } from '@/components/navigation/styled-top-tabs';
 
 export default function MatchsPage() {
   const { t } = useTranslation();
 
-  const { Navigator, Screen } = createMaterialTopTabNavigator();
+  const { Screen } = createMaterialTopTabNavigator();
 
   const listeMatchs = useSelector(
     (state: any) => state.gestionMatchs.listematchs,
@@ -76,7 +77,7 @@ export default function MatchsPage() {
 
     return (
       <HStack className="items-center">
-        <Text className="text-white text-lg mr-2">{title}</Text>
+        <Text className="text-typography-white text-lg mr-2">{title}</Text>
         <FontAwesome5 name={iconName} size={20} color={iconColor} />
         <Text className={`${textColor} ml-1`} size="xl">
           {matchsRestant.toString()}
@@ -86,15 +87,15 @@ export default function MatchsPage() {
   };
 
   return (
-    <Navigator
+    <StyledTopTabs
       screenOptions={{
         title: t('liste_matchs_navigation_title'),
         tabBarScrollEnabled: true,
-        tabBarStyle: { backgroundColor: '#0594ae' },
-        tabBarIndicatorStyle: { backgroundColor: 'white' },
       }}
+      tabBarClassName="bg-custom-background"
+      tabBarIndicatorClassName="bg-custom-bg-inverse"
     >
       {TopTabScreens()}
-    </Navigator>
+    </StyledTopTabs>
   );
 }