+++ src/app/tournoi/(tabs)/resultats.tsx
@@ -27,17 +27,19 @@ const ListeResultats = () => {
   );
 
   return (
-    <VStack className="flex-1 bg-[#0594ae]">
+    <VStack className="flex-1 bg-custom-background">
       <VStack className="flex-1 justify-between">
         <HStack className="flex px-2">
-          <Text className="basis-2/5 text-white text-lg">{t('place')}</Text>
-          <Text className="basis-1/5 text-center text-white text-lg">
+          <Text className="basis-2/5 text-typography-white text-lg">
+            {t('place')}
+          </Text>
+          <Text className="basis-1/5 text-center text-typography-white text-lg">
             {t('victoire')}
           </Text>
-          <Text className="basis-1/5 text-center text-white text-lg">
+          <Text className="basis-1/5 text-center text-typography-white text-lg">
             {t('m_j')}
           </Text>
-          <Text className="basis-1/5 text-right text-white text-lg">
+          <Text className="basis-1/5 text-right text-typography-white text-lg">
             {t('point')}
           </Text>
         </HStack>